# docs/swagger.yaml
basePath: /
definitions:
  accessgrants.Scope:
    enum:
    - pet:read
    - pet:edit_profile
    - events:read
    - events:create
    - events:void
    - attachments:add
    type: string
    x-enum-varnames:
    - ScopePetRead
    - ScopePetEditProfile
    - ScopeEventsRead
    - ScopeEventsCreate
    - ScopeEventsVoid
    - ScopeAttachmentsAdd
  accessgrants.Status:
    enum:
    - invited
    - active
    - revoked
    type: string
    x-enum-varnames:
    - StatusInvited
    - StatusActive
    - StatusRevoked
...
  pets.Sex:
    enum:
    - male
    - female
    - unknown
    type: string
    x-enum-varnames:
    - SexMale
    - SexFemale
    - SexUnknown
  pets.Species:
    enum:
    - dog
    - cat
    type: string
    x-enum-varnames:
    - SpeciesDog
    - SpeciesCat
  pets.createPetRequest:
    properties:
      birth_date:
        description: YYYY-MM-DD opcional
        type: string
      breed:
        description: 'Ej para dog: labrador, poodle. Ej para cat: persian, common.'
        type: string
      name:
        type: string
      notes:
        type: string
      sex:
        allOf:
        - $ref: '#/definitions/pets.Sex'
        enum:
        - male
        - female
        - unknown
      species:
        allOf:
        - $ref: '#/definitions/pets.Species'
        description: dog, cat
        enum:
        - dog
        - cat
    type: object
  pets.grantMini:
    properties:
      id:
        type: string
      status:
        $ref: '#/definitions/accessgrants.Status'
    type: object
  pets.petResponse:
    properties:
      birth_date:
        type: string
      breed:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      notes:
        type: string
      owner_user_id:
        type: string
      sex:
        $ref: '#/definitions/pets.Sex'
      species:
        $ref: '#/definitions/pets.Species'
      updated_at:
        type: string
    type: object
  pets.sharedPetResponse:
    properties:
      grant:
        $ref: '#/definitions/pets.grantMini'
      pet:
        $ref: '#/definitions/pets.petResponse'
      scopes:
        items:
          $ref: '#/definitions/accessgrants.Scope'
        type: array
    type: object
  pets.updatePetRequest:
    properties:
      breed:
        type: string
      name:
        type: string
      notes:
        type: string
      sex:
        allOf:
        - $ref: '#/definitions/pets.Sex'
        enum:
        - male
        - female
        - unknown
      species:
        allOf:
        - $ref: '#/definitions/pets.Species'
        enum:
        - dog
        - cat
    type: object
host: localhost:8080
info:
  contact:
    email: support@odyssoft.com
    name: API Support
    url: https://odyssoft.com/support
  description: API REST para gestión de historiales clínicos de mascotas con sistema
    de delegación de acceso.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Pet Clinical History API
  version: "1.0"
paths:
...
  /me/pets:
    get:
      description: 'Lista las mascotas compartidas con el usuario autenticado mediante
        grants activos que incluyan el scope `pet:read`. Autenticación: `X-Debug-User-ID`
        (dev) o `Authorization: Bearer <token>` (prod).'
      parameters:
      - description: Solo en modo dev, ID de usuario para depuración
        in: header
        name: X-Debug-User-ID
        type: string
      - description: Bearer token en producción
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/pets.sharedPetResponse'
            type: array
      summary: Listar mascotas compartidas conmigo
      tags:
      - pets
  /pets:
    get:
      description: 'Lista todas las mascotas cuyo propietario es el usuario autenticado.
        Autenticación: `X-Debug-User-ID` (dev) o `Authorization: Bearer <token>` (prod).
        Solo el owner ve este listado; los delegados no listan aquí.'
      parameters:
      - description: Solo en modo dev, ID de usuario para depuración
        in: header
        name: X-Debug-User-ID
        type: string
      - description: Bearer token en producción
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/pets.petResponse'
            type: array
        "401":
          description: unauthorized
          schema:
            type: string
      summary: Listar mis mascotas
      tags:
      - pets
    post:
      consumes:
      - application/json
      description: 'Crea una mascota. Species: `dog`, `cat`. Breeds Perro: `labrador`,
        `golden_retriever`, `poodle`, etc. Breeds Gato: `persian`, `common`, etc.'
      parameters:
      - description: Solo en modo dev, ID de usuario para depuración
        in: header
        name: X-Debug-User-ID
        type: string
      - description: Bearer token en producción
        in: header
        name: Authorization
        type: string
      - description: Datos de la mascota; birth_date opcional (YYYY-MM-DD)
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/pets.createPetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/pets.petResponse'
        "401":
          description: unauthorized
          schema:
            type: string
      summary: Crear una mascota
      tags:
      - pets
  /pets/{petID}:
    get:
      description: 'Obtiene el perfil de una mascota. El dueño siempre tiene acceso
        (bypass). Un delegado necesita un grant activo con scope `pet:read`. Autenticación:
        `X-Debug-User-ID` (dev) o `Authorization: Bearer <token>` (prod).'
      parameters:
      - description: Solo en modo dev, ID de usuario para depuración
        in: header
        name: X-Debug-User-ID
        type: string
      - description: Bearer token en producción
        in: header
        name: Authorization
        type: string
      - description: ID de la mascota
        in: path
        name: petID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pets.petResponse'
        "403":
          description: forbidden
          schema:
            type: string
        "404":
          description: pet not found
          schema:
            type: string
      summary: Obtener perfil de mascota
      tags:
      - pets
    patch:
      consumes:
      - application/json
      description: 'Actualiza parcialmente el perfil de una mascota. El dueño siempre
        tiene acceso (bypass). Un delegado necesita un grant activo con scope `pet:edit_profile`.
        Autenticación: `X-Debug-User-ID` (dev) o `Authorization: Bearer <token>` (prod).
        Campo `birth_date` se maneja con semántica PATCH especial: si no se envía,
        no cambia; si se envía como `null`, se limpia; si se envía como string `YYYY-MM-DD`,
        se actualiza.'
      parameters:
      - description: Solo en modo dev, ID de usuario para depuración
        in: header
        name: X-Debug-User-ID
        type: string
      - description: Bearer token en producción
        in: header
        name: Authorization
        type: string
      - description: ID de la mascota
        in: path
        name: petID
        required: true
        type: string
      - description: Campos a actualizar
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/pets.updatePetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pets.petResponse'
        "403":
          description: forbidden
          schema:
            type: string
      summary: Actualizar perfil de mascota
      tags:
      - pets
...
securityDefinitions:
  BearerAuth:
    description: 'Token JWT obtenido de Odin-IAM. Formato: `Bearer <token>`'
    in: header
    name: Authorization
    type: apiKey
  DebugUserID:
    description: Solo en modo dev. Permite simular autenticación sin token real.
    in: header
    name: X-Debug-User-ID
    type: apiKey
swagger: "2.0"